Refactor Validation-Logic
The validation logic and conversions to a GeneratorElement are quite complex and deserve their own place.
