+++ ksp/src/main/kotlin/com/qualityminds/lazyval/ksp/internal/LazyvalKspEnvironment.kt
@@ -1,15 +1,14 @@
-@file:Suppress("TooManyFunctions")
-
 package com.qualityminds.lazyval.ksp.internal
 
 import com.google.devtools.ksp.processing.KSPLogger
 import com.google.devtools.ksp.processing.Resolver
 import com.google.devtools.ksp.processing.SymbolProcessorEnvironment
-import com.google.devtools.ksp.symbol.*
+import com.google.devtools.ksp.symbol.KSClassDeclaration
+import com.google.devtools.ksp.symbol.KSNode
+import com.google.devtools.ksp.symbol.KSType
 import com.qualityminds.lazyval.LazyvalConfiguration
 import com.qualityminds.lazyval.ksp.spi.Generator
 import com.qualityminds.lazyval.ksp.spi.ValidatedKspGeneratorElement
-import com.qualityminds.lazyval.ksp.spi.WrappedProperty
 import java.util.*
 
 internal class LazyvalKspEnvironment(
@@ -21,10 +20,6 @@ internal class LazyvalKspEnvironment(
         const val DISABLED_GENERATORS: String = "lazyval.generators.disable"
         const val BASE_PACKAGE: String = "lazyval.generators.basePackage"
         private const val NO_GENERATION_WARNING = "None of the required classes are available on the classpath! Lazyval will not generate any sources."
-        private const val NOT_FINAL_CLASS_WARNING = "Value Types should not be extendable, hence the class should be final."
-        private const val NOT_FINAL_VALUE_WARNING = "Value Types should be immutable, hence the wrapped property should be final (val)."
-
-
     }
 
     private val logger: KSPLogger = environment.logger
@@ -145,168 +140,4 @@ internal class LazyvalKspEnvironment(
     }
 
 
-    fun validateElement(classDeclaration: KSClassDeclaration): ValidatedKspGeneratorElement? {
-        // KSPs validation is to strict, it won't allow the Java Annotation on a Kotlin class
-//        if (!classDeclaration.validate()) {
-//            environment.logger.error("Invalid Kotlin class declaration", classDeclaration)
-//            return null
-//        }
-
-        return when (classDeclaration.classKind) {
-            ClassKind.CLASS -> validateClass(classDeclaration)
-            else -> {
-                error(classDeclaration, "Only classes and data classes are supported by Lazyval.")
-                null
-            }
-        }
-    }
-
-    private fun validateClass(classDeclaration: KSClassDeclaration): ValidatedKspGeneratorElement? {
-        var valid = true
-
-        if (Modifier.ABSTRACT in classDeclaration.modifiers) {
-            error(classDeclaration, "Abstract class is not a valid ValueType.")
-            valid = false
-        }
-
-        if (Modifier.VALUE in classDeclaration.modifiers) {
-            error(classDeclaration, "value class is not supported by Lazyval.")
-            valid = false
-        }
-
-        // Find properties or their corresponding accessor methods.
-        // Static and @Transient properties are excluded — only the storage payload counts.
-        val publicProperties = classDeclaration.getAllProperties()
-            .filter { !it.isStatic() && !it.isTransient() }
-            .toList()
-        val propertyAccessorPairs = findPropertyAccessorPairs(classDeclaration)
-
-        if (propertyAccessorPairs.size > 1 || publicProperties.size > 1) {
-            error(classDeclaration, "Not a simple ValueType. Lazyval only supports classes with one non-transient property.")
-            valid = false
-        } else if (propertyAccessorPairs.isEmpty() && publicProperties.isEmpty()) {
-            error(classDeclaration, "No accessible properties found. Lazyval requires the ValueType to have exactly one accessible property.")
-            return null
-        }
-
-        if(publicProperties.first().type.resolve().isMarkedNullable) {
-            error(publicProperties.first(), "Wrapped type must not be nullable. Please use a non-nullable type.")
-            valid = false
-        }
-
-        val valueProperty = publicProperties.firstOrNull() ?: propertyAccessorPairs.first().first
-        val accessorMethod = propertyAccessorPairs.firstOrNull()?.second
-
-        // Find factory methods in a companion object
-        val factoryMethods = findFactoryMethods(classDeclaration, valueProperty.type.resolve())
-        if (factoryMethods.size > 1) {
-            val functionNames = factoryMethods.joinToString(", ") { it.simpleName.asString() }
-            error(classDeclaration, "Multiple matching factory methods with the same signature found. Please check functions $functionNames")
-            valid = false
-        }
-
-        val factoryMethod = factoryMethods.firstOrNull()
-
-        return if (valid) {
-            if (Modifier.OPEN in classDeclaration.modifiers) {
-                warn(classDeclaration, NOT_FINAL_CLASS_WARNING)
-            }
-            if (valueProperty.isMutable) {
-                warn(valueProperty, NOT_FINAL_VALUE_WARNING)
-            }
-            ValidatedKspGeneratorElement(
-                classDeclaration,
-                WrappedProperty(valueProperty),
-                factoryMethod,
-                accessorMethod)
-        } else {
-            null
-        }
-    }
-
-    private fun findPropertyAccessorPairs(classDeclaration: KSClassDeclaration): List<Pair<KSPropertyDeclaration, KSFunctionDeclaration?>> {
-        val properties = classDeclaration.getAllProperties()
-            .filter { property ->
-                !property.isStatic() &&
-                        !property.isTransient() &&
-                        property.getter != null &&
-                        property.hasBackingField
-            }
-            .toList()
-
-        val methods = classDeclaration.getAllFunctions()
-            .filter { function ->
-                !function.isStatic() &&
-                        function.parameters.isEmpty() &&
-                        function.returnType != null &&
-                        function.returnType!!.resolve().toString() != "kotlin.Unit"
-            }
-            .toList()
-
-        return properties.mapNotNull { property ->
-            // Look for accessor method that matches the property type
-            val accessor = methods.firstOrNull { method ->
-                method.returnType?.resolve() == property.type.resolve() &&
-                        (method.simpleName.asString() == property.simpleName.asString() ||
-                                method.simpleName.asString() == "${property.simpleName.asString()}()")
-            }
-
-            // For data classes, the accessor might be the property itself
-            // For regular classes an explicit accessor method is needed
-            if (accessor != null || (classDeclaration.modifiers.contains(Modifier.DATA) && !property.isPrivate())) {
-                Pair(property, accessor)
-            } else {
-                null
-            }
-        }
-    }
-
-    private fun findFactoryMethods(classDeclaration: KSClassDeclaration, wrappedType: KSType): List<KSFunctionDeclaration> {
-        val companionObject = classDeclaration.declarations
-            .filterIsInstance<KSClassDeclaration>()
-            .firstOrNull { it.isCompanionObject }
-            ?: return emptyList()
-
-        return companionObject.declarations
-            .filterIsInstance<KSFunctionDeclaration>()
-            .filter { function ->
-                // Check basic structure first
-                if (function.parameters.size != 1) return@filter false
-
-                // Check if return type matches (including nullable variants)
-                val returnType = function.returnType?.resolve() ?: return@filter false
-                val returnTypeMatches = returnType == classDeclaration.asStarProjectedType() ||
-                        (returnType.isMarkedNullable && returnType.makeNotNullable() == classDeclaration.asStarProjectedType())
-                if (!returnTypeMatches) return@filter false
-
-                // Check if parameter type matches (including nullable variants)
-                val paramType = function.parameters[0].type.resolve()
-                val paramTypeMatches = paramType == wrappedType ||
-                        (paramType.isMarkedNullable && paramType.makeNotNullable() == wrappedType)
-
-                paramTypeMatches
-            }
-            .toList()
-    }
-
-    private fun KSPropertyDeclaration.isStatic(): Boolean {
-        return Modifier.CONST in modifiers ||
-                parent is KSClassDeclaration && (parent as KSClassDeclaration).isCompanionObject
-    }
-
-    private fun KSFunctionDeclaration.isStatic(): Boolean {
-        return parent is KSClassDeclaration && (parent as KSClassDeclaration).isCompanionObject
-    }
-
-    private fun KSPropertyDeclaration.isPrivate(): Boolean {
-        return Modifier.PRIVATE in modifiers
-    }
-
-    private fun KSPropertyDeclaration.isTransient(): Boolean {
-        return annotations.any { annotation ->
-            if (annotation.shortName.asString() != "Transient") return@any false
-            val fqn = annotation.annotationType.resolve().declaration.qualifiedName?.asString()
-            fqn == "kotlin.jvm.Transient" || fqn == "java.beans.Transient"
-        }
-    }
 }
